fix: Move page to standard folder, add explanatory line
