+++ src/redesign/components/PolicyEngine.jsx
@@ -20,7 +20,7 @@ import Header from "./Header";
 import Testimonials from "./Testimonials";
 import CalculatorInterstitial from "./CalculatorInterstitial";
 import CitizensEconomicCouncil from "./CitizensEconomicCouncil";
-import APIDocumentationPage from "./APIDocumentationPage";
+import APIDocumentationPage from "../../pages/APIDocumentationPage";
 import CookieConsent from "layout/CookieConsent";
 import TrafwaCalculator from "./TrafwaCalculator";
